+++ src/test/kotlin/com/github/yusufugurozbek/testcontainers/port/updater/constants/MongoDbTestConstants.kt
@@ -10,4 +10,11 @@ object MongoDbTestConstants {
     const val MONGODB_LOG_ENTRY =
         "00:34:59.184 [main] DEBUG com.github.yusufugurozbek.testcontainersdemo.MongoDbTest" +
                 " - Database: mongodb://localhost:54321/test"
+    const val MONGODB_JSON_LOG_ENTRY = "{\"@timestamp\":\"2021-07-16T15:11:23.4112+02:00\"," +
+            "\"@version\":\"1\"," +
+            "\"message\":\"Database: mongodb://localhost:54321/test\"," +
+            "\"logger_name\":\"org.flywaydb.core.internal.database.base.BaseDatabaseType\"," +
+            "\"thread_name\":\"main\"," +
+            "\"level\":\"INFO\"," +
+            "\"level_value\":20000}"
 }