+++ src/test/kotlin/com/github/yusufugurozbek/testcontainers/port/updater/constants/MsSqlTestConstants.kt
@@ -21,4 +21,14 @@ object MsSqlTestConstants {
             "connectRetryCount=1;maxResultBuffer=-1;" +
             "applicationName=Microsoft JDBC Driver for SQL Server;" +
             "applicationIntent=readwrite; (Microsoft SQL Server 14.0)"
+    const val MSSQL_JSON_LOG_ENTRY = "{\"@timestamp\":\"2021-07-16T15:11:23.4112+02:00\"," +
+            "\"@version\":\"1\"," +
+            "\"message\":\"Database: jdbc:sqlserver://localhost:54321;connectRetryInterval=10;" +
+            "connectRetryCount=1;maxResultBuffer=-1;" +
+            "applicationName=Microsoft JDBC Driver for SQL Server;" +
+            "applicationIntent=readwrite; (Microsoft SQL Server 14.0)\"," +
+            "\"logger_name\":\"org.flywaydb.core.internal.database.base.BaseDatabaseType\"," +
+            "\"thread_name\":\"main\"," +
+            "\"level\":\"INFO\"," +
+            "\"level_value\":20000}"
 }